Added edit for each model/content
Roots to content#index.
Content#index shows all the models available for edit in a dropdown.
Once a model is selected in the dd, an edit iframe enables editing of
the model's content and style via a ckeditor textarea.
// unfinished : /uploads shows a screen where one can upload via drag
and drop or simple file select an image to upload to server via
carrierwave and dropzonejs.
// unfinished : added cloudinary to upload image to cloud